refactor(contracts): rename file name
ZSwap and ZTransaction file name casing has been changed

Issue-URL: https://app.clickup.com/t/8695rpf2h

+++ contracts/contracts/protocol/v1/core/libraries/NullifierSpender.sol
@@ -11,7 +11,7 @@ library NullifierSpender {
         mapping(bytes32 => uint256) storage isSpent,
         uint256[3] memory nullifiers
     ) internal {
-        for (uint nullifier = 0; nullifier < nullifiers.length; ) {
+        for (uint256 nullifier = 0; nullifier < nullifiers.length; ) {
             validateAndSpendNullifier(isSpent, nullifiers[nullifier]);
             unchecked {
                 ++nullifier;
